+++ mocktracer/src/utility.h
@@ -1,7 +1,7 @@
 #ifndef OPENTRACING_UTILITY_H
 #define OPENTRACING_UTILITY_H
 
-#include <opentracing/mock_tracer.h>
+#include <opentracing/mocktracer/tracer.h>
 
 namespace opentracing {
 BEGIN_OPENTRACING_ABI_NAMESPACE
